refactor: decompose MetricsCallback into MetricsCallback and BestMetricTrackerCallback
- Split MetricsCallback into two separate callbacks for better separation of concerns
- MetricsCallback: handles metrics computation and logging for train/val/test stages
- BestMetricTrackerCallback: tracks best validation metric for model selection
- Updated callback configurations and experiment configs to use both callbacks
- Fixed tracked_metric_name references in default.yaml to point to BestMetricTrackerCallback
- Updated base_module.py comment to reflect new callback structure
- Enhanced documentation in ARCHITECTURE.md with separate sections for each callback
- Added comprehensive docstrings to BestMetricTrackerCallback

This refactoring provides cleaner architecture with single-responsibility callbacks.
