+++ configs/callbacks/default.yaml
@@ -4,20 +4,21 @@ defaults:
   - model_summary
   - rich_progress_bar
   - metrics_callback
+  - best_metric_tracker_callback
   - _self_
 
 model_checkpoint:
   dirpath: ${paths.output_dir}/checkpoints
   filename: "epoch_{epoch:03d}"
   monitor: "val/best"
-  mode: ${eval:"'max' if '${oc.select:callbacks.metrics_callback.tracked_metric_name,''}' else 'min'"} # "max" if metric tracked, "min" for loss
+  mode: ${eval:"'max' if '${oc.select:callbacks.best_metric_tracker_callback.tracked_metric_name,''}' else 'min'"} # "max" if metric tracked, "min" for loss
   save_last: True
   auto_insert_metric_name: False
 
 early_stopping:
   monitor: "val/best"
   patience: 100
-  mode: ${eval:"'max' if '${oc.select:callbacks.metrics_callback.tracked_metric_name,''}' else 'min'"} # "max" if metric tracked, "min" for loss
+  mode: ${eval:"'max' if '${oc.select:callbacks.best_metric_tracker_callback.tracked_metric_name,''}' else 'min'"} # "max" if metric tracked, "min" for loss
 
 model_summary:
   max_depth: -1
