+++ configs/callbacks/metrics_callback.yaml
@@ -3,4 +3,3 @@
 metrics_callback:
   _target_: ml_core.callbacks.MetricsCallback
   metrics: null # metrics composition to track, should be specified in experiment config
-  tracked_metric_name: null # metric to track for best model selection, if null tracks validation loss
