+++ ml_core/callbacks/__init__.py
@@ -1,5 +1,6 @@
 """Callbacks for PyTorch Lightning training."""
 
+from ml_core.callbacks.best_metric_tracker_callback import BestMetricTrackerCallback
 from ml_core.callbacks.metrics_callback import MetricsCallback
 
-__all__ = ["MetricsCallback"]
+__all__ = ["MetricsCallback", "BestMetricTrackerCallback"]
