+++ ml_core/callbacks/metrics_callback.py
@@ -1,7 +1,6 @@
 from typing import Any, Literal
 
 from lightning import Callback, LightningModule, Trainer
-from torchmetrics import MaxMetric, MinMetric
 
 from ml_core.models.utils import MetricsComposition
 
@@ -11,24 +10,19 @@ class MetricsCallback(Callback):
 
     This callback creates metric collections for each stage and attaches them to the
     LightningModule, then updates and logs them during training. Metrics are kept as module
-    attributes to follow Lightning's expected pattern. Also tracks the best validation metric for
-    model selection.
+    attributes to follow Lightning's expected pattern.
     """
 
     def __init__(
         self,
         metrics: MetricsComposition | None = None,
-        tracked_metric_name: str | None = None,
     ) -> None:
         """Initialize the metrics callback.
 
         :param metrics: Optional composition of metrics to track across stages.
-        :param tracked_metric_name: Metric key (without stage prefix) used to track the best value
-            on validation; if None, the total validation loss is tracked.
         """
         super().__init__()
         self.metrics = metrics
-        self.tracked_metric_name = tracked_metric_name
 
     def setup(self, trainer: Trainer, pl_module: LightningModule, stage: str) -> None:
         """Create and attach metric collections to the module before training starts.
@@ -37,43 +31,11 @@ class MetricsCallback(Callback):
         :param pl_module: The Lightning module to attach metrics to.
         :param stage: Current stage (fit, test, validate, predict).
         """
-        # Set tracked_metric_name on the module so it can be used for scheduler config
-        if not hasattr(pl_module, "tracked_metric_name"):
-            pl_module.tracked_metric_name = self.tracked_metric_name
-
         # Only create metrics once, and only if they don't already exist
         if self.metrics is not None and not hasattr(pl_module, "train_metrics"):
             pl_module.train_metrics = self.metrics.clone("train/")
             pl_module.val_metrics = self.metrics.clone("val/")
             pl_module.test_metrics = self.metrics.clone("test/")
-
-        # Create best validation metric tracker
-        # Use MaxMetric for custom metrics (higher is better), MinMetric for loss (lower is better)
-        if not hasattr(pl_module, "best_val_tracked_metric"):
-            pl_module.best_val_tracked_metric = (
-                MaxMetric() if self.tracked_metric_name else MinMetric()
-            )
-
-    def on_train_start(self, trainer: Trainer, pl_module: LightningModule) -> None:
-        """Reset validation metrics at the start of training.
-
-        Note: We reset best_val_tracked_metric only when starting fresh training,
-        not when resuming from a checkpoint. When resuming, it gets loaded automatically.
-
-        :param trainer: The Lightning trainer.
-        :param pl_module: The Lightning module.
-        """
-        if hasattr(pl_module, "val_metrics"):
-            for metric_name in pl_module.val_metrics:
-                pl_module.val_metrics[metric_name].reset()
-
-        # Reset best metric tracker only if NOT resuming from checkpoint
-        # (trainer.ckpt_path is set when resuming)
-        if hasattr(pl_module, "best_val_tracked_metric"):
-            if trainer.ckpt_path is None:
-                # Fresh training - reset the tracker for validation sanity checks
-                pl_module.best_val_tracked_metric.reset()
-            # else: resuming from checkpoint - don't reset, it was loaded from checkpoint
 
     def _on_stage_batch_end(
         self,
@@ -158,29 +120,3 @@ class MetricsCallback(Callback):
         :param dataloader_idx: Index of the current dataloader.
         """
         self._on_stage_batch_end("test", pl_module, outputs)
-
-    def on_validation_epoch_end(self, trainer: Trainer, pl_module: LightningModule) -> None:
-        """Compute and log best validation metric at epoch end.
-
-        :param trainer: The Lightning trainer.
-        :param pl_module: The Lightning module.
-        """
-        if self.tracked_metric_name is None:
-            tracked_metric = pl_module.val_losses["total"]
-        else:
-            if not hasattr(pl_module, "val_metrics"):
-                raise ValueError(
-                    f"tracked_metric_name '{self.tracked_metric_name}' is set but "
-                    "val_metrics not found. Did you set metrics in MetricsCallback?"
-                )
-            tracked_metric = pl_module.val_metrics[self.tracked_metric_name]
-
-        pl_module.best_val_tracked_metric(tracked_metric.compute())
-        pl_module.log(
-            "val/best",
-            pl_module.best_val_tracked_metric.compute(),
-            on_step=False,
-            on_epoch=True,
-            prog_bar=True,
-            sync_dist=True,
-        )
